+++ express/src/decorators/controller.ts
@@ -1,17 +1,17 @@
-import { ExpressMeta } from '../interface';
-import { getMeta, getMiddleware } from '../meta';
+import { ExpressMeta, getMeta } from '../meta';
+import { Middleware } from '../middleware';
 
 /**
  * Registers controller for base url
- * @param {string} baseUrl
- * @param {Function|Function[]} [middleware]
- * @returns {(target:Function)=>void}
- * @constructor
+ *
+ * @param {string} url
+ * @param {Middleware[]} [middleware]
  */
-export let Controller = (baseUrl: string, middleware?: Function|Function[]): ClassDecorator  => {
-  return (target: Function): void => {
-    let meta: ExpressMeta = getMeta(target.prototype);
-    meta.baseUrl = baseUrl;
-    meta.controllerMiddleware = getMiddleware(middleware);
+export function Controller(url: string, middleware?: Middleware[]) {
+  return (target): void => {
+    const meta: ExpressMeta = getMeta(target.prototype);
+
+    meta.url = url;
+    meta.middleware = middleware;
   };
 };