+++ express/src/express.ts
@@ -1,89 +1,161 @@
-import { Router, Express, RequestHandler } from 'express';
+import { RequestHandler, ErrorRequestHandler, Application, Router, Express, Request, Response, NextFunction } from 'express';
+import { Container, InjectionToken } from '@decorators/di';
 
-import { ParameterType, Routes, Middleware, Injectable, ExpressClass, Params } from './interface';
+import { ExpressMeta, getMeta, ParameterType, ExpressClass, Route, ParameterConfiguration } from './meta';
+import { middlewareHandler, ErrorMiddleware } from './middleware';
 
-function getParam(source: any, paramType: string, name: string) {
-  let param = source[paramType] || source;
-  return param[name] || param;
+/**
+ * Error Middleware class registration DI token
+ */
+export const ERROR_MIDDLEWARE = new InjectionToken('ERROR_MIDDLEWARE');
+
+/**
+ * Attach controllers to express application
+ *
+ * @param {Express} app Express application
+ * @param {ExpressClass[]} controllers Controllers array
+ */
+export function attachControllers(app: Express, controllers: ExpressClass[]) {
+  controllers.forEach((controller: ExpressClass) => registerController(app, controller));
+
+  app.use(errorMiddlewareHandler());
 }
 
-function extractParameters(req, res, next, params): any[] {
-  let args = [];
-  if (!params || !params.length) {
-    return [req, res, next];
-  }
-  for (let item of params) {
+/**
+ * Register controller via registering new Router
+ *
+ * @param {Application} app
+ * @param {ExpressClass} Controller
+ * @returns
+ */
+function registerController(app: Application, Controller: ExpressClass) {
+  const controller: ExpressClass = Container.get(Controller);
+  const meta: ExpressMeta = getMeta(controller);
+  const router: Router = Router();
 
-    switch (item.type) {
-      default: args[item.index] = res; break; // response
-      case ParameterType.REQUEST: args[item.index] = getParam(req, null, item.name); break;
-      case ParameterType.NEXT: args[item.index] = next; break;
-      case ParameterType.PARAMS: args[item.index] = getParam(req, 'params', item.name); break;
-      case ParameterType.QUERY: args[item.index] = getParam(req, 'query', item.name); break;
-      case ParameterType.BODY: args[item.index] = getParam(req, 'body', item.name); break;
-      case ParameterType.HEADERS: args[item.index] = getParam(req, 'headers', item.name); break;
-      case ParameterType.COOKIES: args[item.index] = getParam(req, 'cookies', item.name); break;
-    }
+  const routes: object = meta.routes;
+  const url: string = meta.url;
+  const params: object = meta.params;
 
-  }
-  return args;
-}
+  /**
+   * Wrap all registered middleware with helper function
+   * that can instantiate or get from the container instance of the class
+   * or execute given middleware function
+   * @see getMiddleware
+   */
+  const routerMiddleware: RequestHandler[] = (meta.middleware || [])
+    .map(middleware => middlewareHandler(middleware));
 
-function registerController(app, Controller, deps) {
-  let controller: ExpressClass = new Controller(...deps);
-  let router: Router = Router();
-  let controllerMiddleware: RequestHandler[] = controller.__express_meta__.controllerMiddleware;
-  let routes: Routes = controller.__express_meta__.routes;
-  let middleware: Middleware = controller.__express_meta__.middleware;
-  let baseUrl: string = controller.__express_meta__.baseUrl;
-  let params: Params = controller.__express_meta__.params;
-
-  if (controllerMiddleware.length) {
-    router.use(...controllerMiddleware);
+  /**
+   * Apply router middleware
+   */
+  if (routerMiddleware.length) {
+    router.use(...routerMiddleware);
   }
 
+  /**
+   * Applying registered routes
+   */
   for (const methodName of Object.keys(routes)) {
-    let method: string = routes[methodName].method;
-    let fn: Function;
+    const route: Route = routes[methodName];
 
-    fn = (req, res, next) => {
-      let args = extractParameters(req, res, next, params[methodName]);
+    const routeHandler = (req, res, next) => {
+      const args = extractParameters(req, res, next, params[methodName]);
+
       return controller[methodName].apply(controller, args);
     };
 
-    let routeArgs = [
-      routes[methodName].url, ...(middleware[methodName] || []), fn
-    ];
+    const routeMiddleware: RequestHandler[] = (route.middleware || [])
+      .map(middleware => middlewareHandler(middleware));
 
-    router[method].apply(router, routeArgs);
+    router[route.method].apply(router, [
+      route.url, ...routeMiddleware, routeHandler
+    ]);
   }
 
-  app.use(baseUrl, router);
+  app.use(url, router);
 
   return app;
 }
 
 /**
- * Attach controllers to express application
- * @param {Express} app Express application
- * @param {Controller} controllers Controllers array
+ * Add error middleware to the app
+ *
+ * @param {Express} app
  */
-export function attachControllers(app: Express, injectables: Array<Injectable | ExpressClass>) {
-  try {
-    injectables
-      .forEach((injectable: Injectable | ExpressClass) => {
-        const controller = (<Injectable>injectable).provide || <ExpressClass>injectable;
-        const deps = (<Injectable>injectable).deps || [];
-
-        registerController(app, controller, deps);
-      });
-  } catch (e) {
-    console.log('Second parameter should be array of injectables: { provide: Controller, deps: any[] }', e, e.message);
+function errorMiddlewareHandler(): ErrorRequestHandler {
+  return function(error: Error, req: Request, res: Response, next: NextFunction): void {
+    try {
+      const errorMiddleware: ErrorMiddleware = Container.get(ERROR_MIDDLEWARE);
+      errorMiddleware.use(error, req, res, next);
+    } catch (error) {
+      next(error);
+    }
   }
 }
 
 /**
- * @deprecated
- * @see attachControllers
+ * Extract parameters for handlers
+ *
+ * @param {Request} req
+ * @param {Response} res
+ * @param {NextFunction} next
+ * @param {ParameterConfiguration[]} params
+ *
+ * @returns {any[]}
  */
-export let bootstrapControllers = attachControllers;
+function extractParameters(req: Request, res: Response, next: NextFunction, params: ParameterConfiguration[]): any[] {
+  if (!params || !params.length) {
+    return [ req, res, next ];
+  }
+
+  const args = [];
+
+  for (let { name, index, type } of params) {
+
+    switch (type) {
+      case ParameterType.RESPONSE:
+        args[index] = res;
+        break;
+      case ParameterType.REQUEST:
+        args[index] = getParam(req, null, name);
+        break;
+      case ParameterType.NEXT:
+        args[index] = next;
+        break;
+      case ParameterType.PARAMS:
+        args[index] = getParam(req, 'params', name);
+        break;
+      case ParameterType.QUERY:
+        args[index] = getParam(req, 'query', name);
+        break;
+      case ParameterType.BODY:
+        args[index] = getParam(req, 'body', name);
+        break;
+      case ParameterType.HEADERS:
+        args[index] = getParam(req, 'headers', name);
+        break;
+      case ParameterType.COOKIES:
+        args[index] = getParam(req, 'cookies', name);
+        break;
+    }
+
+  }
+
+  return args;
+}
+
+/**
+ * Get parameter value from the source object
+ *
+ * @param {*} source
+ * @param {string} paramType
+ * @param {string} name
+ *
+ * @returns {*}
+ */
+function getParam(source: any, paramType: string, name: string): any {
+  let param = source[paramType] || source;
+
+  return param[name] || param;
+}