+++ express/CHANGELOG.md
@@ -1,3 +1,13 @@
+# Express#2.0.0
+* Dependency injection with `@decorators/di`, closes #54;
+* Express route middlewares as a class, closes #55; 
+* Functions as middleware are no longer supported, use classes instead
+* Single middleware are no longer supported, use array instead (unified interface for all usages)
+* Added `All`, `Patch` and `Head` route decorators
+* Updated `express` version to `>=4.16.2`
+* Removed deprecated methods and decorators
+* big refactoring
+
 # Express#1.3.0
 * Moved express-related types to package interfaces
 * **@Request(name?: string)** decorator accepts optional parameter name